+++ packages/astro/src/core/session.ts
@@ -6,6 +6,7 @@ import {
 	builtinDrivers,
 	createStorage,
 } from 'unstorage';
+import type { AstroSettings } from '../types/astro.js';
 import type {
 	ResolvedSessionConfig,
 	SessionConfig,
@@ -13,7 +14,13 @@ import type {
 } from '../types/public/config.js';
 import type { AstroCookies } from './cookies/cookies.js';
 import type { AstroCookieSetOptions } from './cookies/cookies.js';
-import { SessionStorageInitError, SessionStorageSaveError } from './errors/errors-data.js';
+import {
+	SessionConfigMissingError,
+	SessionConfigWithoutFlagError,
+	SessionStorageInitError,
+	SessionStorageSaveError,
+	SessionWithoutSupportedAdapterOutputError,
+} from './errors/errors-data.js';
 import { AstroError } from './errors/index.js';
 
 export const PERSIST_SYMBOL = Symbol();
@@ -98,7 +105,11 @@ export class AstroSession<TDriver extends SessionDriverName = any> {
 	/**
 	 * Gets a session value. Returns `undefined` if the session or value does not exist.
 	 */
-	async get<T = any>(key: string): Promise<T | undefined> {
+	async get<T = void, K extends string = string>(
+		key: K,
+	): Promise<
+		(T extends void ? (K extends keyof App.SessionData ? App.SessionData[K] : any) : T) | undefined
+	> {
 		return (await this.#ensureData()).get(key)?.data;
 	}
 
@@ -145,7 +156,15 @@ export class AstroSession<TDriver extends SessionDriverName = any> {
 	 * Sets a session value. The session is created if it does not exist.
 	 */
 
-	set<T = any>(key: string, value: T, { ttl }: { ttl?: number } = {}) {
+	set<T = void, K extends string = string>(
+		key: K,
+		value: T extends void
+			? K extends keyof App.SessionData
+				? App.SessionData[K]
+				: any
+			: NoInfer<T>,
+		{ ttl }: { ttl?: number } = {},
+	) {
 		if (!key) {
 			throw new AstroError({
 				...SessionStorageSaveError,
@@ -462,15 +481,39 @@ export class AstroSession<TDriver extends SessionDriverName = any> {
 	}
 }
 // TODO: make this sync when we drop support for Node < 18.19.0
-export function resolveSessionDriver(driver: string | undefined): Promise<string> | string | null {
+export async function resolveSessionDriver(driver: string | undefined): Promise<string | null> {
 	if (!driver) {
 		return null;
 	}
-	if (driver === 'fs') {
-		return import.meta.resolve(builtinDrivers.fsLite);
+	try {
+		if (driver === 'fs') {
+			return await import.meta.resolve(builtinDrivers.fsLite);
+		}
+		if (driver in builtinDrivers) {
+			return await import.meta.resolve(builtinDrivers[driver as keyof typeof builtinDrivers]);
+		}
+	} catch {
+		return null;
 	}
-	if (driver in builtinDrivers) {
-		return import.meta.resolve(builtinDrivers[driver as keyof typeof builtinDrivers]);
-	}
+
 	return driver;
+}
+
+export function validateSessionConfig(settings: AstroSettings): void {
+	const { experimental, session } = settings.config;
+	const { buildOutput } = settings;
+	let error: AstroError | undefined;
+	if (experimental.session) {
+		if (!session?.driver) {
+			error = new AstroError(SessionConfigMissingError);
+		} else if (buildOutput === 'static') {
+			error = new AstroError(SessionWithoutSupportedAdapterOutputError);
+		}
+	} else if (session?.driver) {
+		error = new AstroError(SessionConfigWithoutFlagError);
+	}
+	if (error) {
+		error.stack = undefined;
+		throw error;
+	}
 }