+++ packages/astro/test/core-image.test.js
@@ -1135,6 +1135,18 @@ describe('astro:image', () => {
 			assert.equal(response.headers.get('content-type'), 'image/webp');
 		});
 
+		it('returns HEAD method ok for /_image', async () => {
+			const params = new URLSearchParams();
+			params.set('href', '/src/assets/penguin1.jpg?origWidth=207&origHeight=243&origFormat=jpg');
+			params.set('f', 'webp');
+			const response = await fixture.fetch('/some-base/_image?' + String(params), {
+				method: 'HEAD',
+			});
+			assert.equal(response.status, 200);
+			assert.equal(response.body, null);
+			assert.equal(response.headers.get('content-type'), 'image/webp');
+		});
+
 		it('does not interfere with query params', async () => {
 			let res = await fixture.fetch('/api?src=image.png');
 			const html = await res.text();