+++ packages/integrations/node/test/fixtures/api-route/astro.config.mjs
@@ -1,0 +1,7 @@
+import { defineConfig} from "astro/config";
+
+export default defineConfig({
+    security: {
+        checkOrigin: false
+    }
+})